RAG API/requirements.txt
# Core API framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# LangChain for RAG pipeline
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28

# Azure services
azure-search-documents==11.5.3
azure-identity==1.23.1

# OpenAI for Azure OpenAI integration
openai==1.98.0

# Environment management
python-dotenv==1.1.1

# Data validation
pydantic==2.11.7

# HTTP requests (for testing)
requests==2.32.4